chore: clean up rubocop warnings
Clean up a number of rubocop warnings throughout the codebase and bump
the required ruby version to 2.7 as some of the changes require at least
ruby 2.6 (and ruby 2.5 is *very* out of date).

Also regenerated the rubocop todo file as some of the violations are not
things I was willing to change at this time. This means we have a known
starting point and we can run rubocop as part of the build to ensure
that further issues are not added in the future.