chore: upgrade build pipeline
Previously we were using buildkite legacy steps to run the specs for a
single version of ruby on the shared `teg-ci` build image.

Migrate the build to a buildkite pipeline.yml file and run the build on
multiple different ruby versions (2.7, 3, 3,1 and 3.2). This approach is
based on that used for the barcodevalidation gem and makes it easier to
manage and extend our build pipeline in the future (as well as moving to
testing on multiple versions of ruby).