chore: migrate pipeline definition to repository
Previously the pipeline was defined in buildkite and the actual build
was in a batch file. To enable us to migrate to a multi-step build using
docker-compose described in pipeline.yml we need to migrate the existing
build definition into the repo before we upgrade it.

We also remove the ecr get-login as that is no longer required.